@@ -7,7 +7,7 @@
         "bin/easy-testing"
     ],
     "require": {
-        "php": ">=7.3",
+        "php": ">=8.0",
         "nette/utils": "^3.2",
         "symfony/finder": "^4.4|^5.3",
         "symfony/http-kernel": "^4.4|^5.3",
