@@ -7,22 +7,12 @@ namespace Symplify\EasyTesting\ValueObject;
 final class InputAndExpected
 {
     /**
-     * @var string
-     */
-    private $input;
-
-    /**
-     * @var mixed
-     */
-    private $expected;
-
-    /**
      * @param mixed $expected
      */
-    public function __construct(string $original, $expected)
-    {
-        $this->input = $original;
-        $this->expected = $expected;
+    public function __construct(
+        private string $input,
+        private $expected
+    ) {
     }
 
     public function getInput(): string
